feat: 增加target指令
